requirements.txt
crewai
crewai-tools
agentops
tavily-python
scrapegraph-py
python-dotenv
google-generativeai